requirements.txt
tensorflow==2.15.0
flask
gunicorn
google-cloud-storage
pillow
numpy